Add a blog
This commit is extremely bloated, but it's the first step in adding an
eleventy blog to this site. It updates a bunch of styling, it's clearly
not finished, there are TODOs everywhere, but not committing something
seemed like a super bad idea. I still have to get the homepage added in
there and figure out a story for any JS I want to build. It'd also be
nice to add some sort of analytics tracking. We'll see.